backend/pyproject.toml
[project]
name = "nmtsa-lms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.5",
    "django>=5.2.7",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.0",
    "django-filter>=23.5",
    "django-taggit>=6.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "pillow>=11.0.0",
    "moviepy>=1.0.3",
    "drf-spectacular>=0.27.0",
]
